Don't do lossy normalization to the cassette data.
The normalization was originally done so that the cassette contents would be the same, regardless of the HTTP library used.  However, this normalization was lossy and could not always be accurately denormalized.

The most important thing is that VCR replays a response exactly as it originally occurred.

For more background on this, see https://github.com/geemus/fog/issues/434.

Closes #78.