Work around weird ruby 1.8 bug.
See https://gist.github.com/1324150 and https://github.com/rspec/rspec-core/issues/487 for examples of the bug.  Because I was calling a method that used defined? on a nested constant in a rescue clause, it caused the interpreter to immediately exit.  This fixes it by not calling the method in the rescue clause.