Remove unnecessary URI normalization in the :uri matcher.
This was just needed to get the test to pass that created a cassette with a particular URI and then tested a request against it.  Different HTTP libraries normalize URIs differently so we had to mirror that normalization as well.

However, if we test the recording and the playback, we don't need to do manual normalization, because we will record the URI in the same form it will be played back since the HTTP library will treat it the same way both times.