Deterministically boot the sinatra app in spec helper.
Before, it would be booted the first time the server was referenced to get the port.  This could lead to failures if the specs were run in a different order or a subset of them were run, though, because the current VCR settings could affect the booting of the server.