Ensure excon status errors are handled properly.
When you make an excon request with the :expects option, it will raise a subclass of Excon::Errors::HTTPStatusError if the response is not the expected status code.  We were manually re-raising this error before, but the excon code up the call stack* rescues all errors and raises an Excon::Errors::SocketError.  This caused the error class to be different then what it would have been otherwise.  Instead, we want to return the raw response (i.e. no expects error), and we let excon transform this in to a status error.

* https://github.com/geemus/excon/blob/v0.9.5/lib/excon/connection.rb#L225-232