Fix excon so real requests are made with a connection constructed with same args.
I got this warning from VCR on some tests using fog/excon to download a file from S3 over SSL:

WARNING: VCR could not extract a response from Excon error (#<OpenSSL::SSL::SSLError: hostname does not match the server certificate>)

I discovered that the SSL cert verification for a real request w/o VCR was made against
`s3.amazonaws.com`, while the verification for the request being recorded by VCR was made
against `<bucket_name>.s3.amazonaws.com`. The difference is due to the fact that fog
constructs an Excon::Connection instance with a URI of `s3.amazonaws.com`, then makes a
request with a :host param set to the FQDN. Before this commit, VCR made the real
request by constructing a new Excon::Connection instance with the URI set to the full
URI, and it caused the SSL cert verification failure.

This fixes it by ensuring the "real request" connection instance is constructed with
the same args as the original connection instance.