Remove separate spec for non-nil effective_url.
We don't need a separate spec for this. We don't really care about
whether or not the effective url is nil or not (typhoeus may set it
to nil in certain situations). It's just an assumption of the other
spec: `played_back.should eq(recorded)` only demonstrates that VCR
has properly set the effective_url if typhoeus set a non-nil value
on that attribute. Otherwise, the spec could pass even w/o VCR
setting the value, due to it defaulting to nil.