Remove failing faraday cukes.
Faraday has some additional headers (see bf9b312a1f6dc2da048d935c2925de85ac1f95c1) that
do not match what is in the VCR cassette for this example. We've got this covered with
some specs already, and there's not much documentation value here to working around it,
so it's simplest just to remove the faraday scenarios.