normalize response charset for Ruby 2.0 in tests
Ruby 2.0 initializes HTTP response bodies as UTF-8 (coincidentally,
Sinatra indicates by default that the responses it sends are UTF-8, but
Net::HTTP doesn't pay attention to that), where older Rubies interpreted
them as US-ASCII and thus VCR serialized it that way. For the purpose of
this test, force the response body to US-ASCII so that the outcome of
serialization is normalized between different Ruby versions.