Revert "travis: cache bundle across jobs"
This reverts commit c5e90d06ad8adb0cc066a04cb22fe22c3cc2935c.

Merging that commit caused the build to start failing due to the fact
that it changed where the bundled gems are installed, and our
warnings prevention assumes a warning originated from VCR (and thus is
worthy of failing the build) if it includes a filename that is in the
current working directory.  Previously the bundled gems were
installed outside of the working directory and thus warnings from
dependent gems (like WebMock) would not trigger a failure.

We should re-work the warnings prevention logic, but for now
I'm reverting this to get back to green.