Work around Faraday 0.8 incompatibility in tests
This test verifies that the exact same response body is played back
after being recorded by VCR. However, Faraday 0.8 may return nil for
response body if it was a HTTP 204. VCR internals then cast that value
into string, since they assume that a body can't be anything else. So
this really is an issue partly with VCR itself.

Now this test is skipped in case of Faraday < 0.9