Randomize test server ports in cukes
Previously, cukes typically used ports 7777 and 8888 for spinning up the
Sinatra test server. This would fail when the user already has something
running on one of these ports, like Charles proxy which typically runs
on 8888.

Now the ports are randomized and picked based on availability. This also
means that there is no chance of port collisions in the condition when
one subprocess is shutting down a server while the other is starting a
new one.