Skip Bundler and Rubygems while executing ruby scripts in cukes
Speeds up subprocess execution by preparing the LOAD_PATH for nested
interpreter invocations. Some attempt to do this existed previously but
it wasn't efficent and did rely on Rubygems nevertheless.

Bundler is only used for subprocesses that need to invoke gem binstubs
such as `cucumber` and `rspec`. Such cukes are tagged "@with-bundler".