Get `uri_parser` test passing on Ruby 2.2
The `uri_parser` test was dependent on the subdomain URI parsing bug in
Ruby's own stdlib, but that bug got fixed in Ruby 2.2. It was also
dependent on Addressable which is not a direct test dependency, but was
instead brought in by WebMock.

Now, the test doesn't rely on either Addressable nor Ruby bug. Instead,
we simply confirm that `uri_parser` will take over URI parsing if it's
defined.