make ignore_cassettes a boolean again to prevent nil.dup crashes
According to the documentation turn_off's insert_cassettes is a boolean, if
it is not set as an option it will however be implicitly set to nil since
the thread-safety port. This in turn makes requests fail on nil.dup after
turn_off was called as dup_context would explicitly call dup for
ignore_cassettes.

To prevent this ignore_cassettes is now used as a boolean everywhere again
and defaults to false when turn_off! is called without ignore_cassettes
explicitly set. This should prevent crashes in dup_context.