Re-record is opt-in; made that clearer
The previous wording implied that re-recording is _automatic_, i.e. enabled by default. The fact is that it isn't, as stated by the Rubydocs for the `use_cassette` method: http://www.rubydoc.info/gems/vcr/VCR#use_cassette-instance_method

The purpose of this PR is to try and make this more apparent here for people like me, who are new to the VCR project.