remove test files from built gem
I got the idea when we got a similar pull request on resque-web.
https://github.com/resque/resque-web/pull/118

End users shouldn't need the test files and they take up space on disk.

The `test_files` method is left over from when `gem test` was a thing.
https://github.com/rubygems/rubygems/commit/429f883210f8b2b38ea310f7fc6636cd0e456d5c

There is some more discussion on this in a rubygems issue:
https://github.com/rubygems/rubygems/issues/735