Handle UnusedHTTPInteractionError while making UnhandledHTTPRequestError
When VCR raises `UnhandledHTTPRequestError`, it intends to terminate exectuion
of the current test case and give the use a helpful error message.
While constructing `UnhandledHTTPRequestError`, it ejects all cassetees
currently in use to collect them. This has the unfortunate side-effect
of possibly raising `UnusedHTTPInteractionError` if
`allow_unused_http_interactions` is set to false.

The `UnusedHTTPInteractionError` from this situation is misleading.
A test case can consume all the requests in a cassette, but if it
does a superfluous unknown request before doing so, it is terminated
with this error.

This commit catches `UnusedHTTPInteractionError` when constructing
`UnhandledHTTPRequestError` to give a less misleading error message.