Mitigate threading issues on the HTTP interactions list
One VCR threading issue is that when multiple request come across in separate threads, the list interactions is not synchronized, causing errors due to loss of tracking of the un/used interactions.

This commit places two mutexes in the core sections of contention:

- the first instantiation of HTTPInteractionList, which without mutex, yields to a second instance overwriting the first (reference)
- the transfer of a given request from the unused interactions list to the used one, and other problems

Please note that this commit is not intended to make VCR thread safe in the absolute sense, but it intends to mitigate (possibly not solve) one specific threading problem experienced by multiple users.

Closes #200, depending on the maintaners intention towards VCR thread-safety.