Fix broken build due to Hashdiff deprecation
webmock uses the hashdiff gem, which recently introduced a deprecation
warning regarding the spelling of the `Hashdiff` constant.

```
The HashDiff constant used by this gem conflicts with another gem of a similar name.
As of version 1.0 the HashDiff constant will be completely removed and replaced by Hashdiff.
```

This warning in the output causes one of the vcr cucumber tests to fail.

The fix is to update to the latest version of the hashdiff gem, which is
currently in beta.