Fix turning off VCR not working when server thread already started
When VCR is first used in the main (test) thread, it initializes a
context, holding the status of whether VCR is turned off. When another
thread is started (ie. app server), VCR initializes a context there by
duplicating the main thread's context and linking to the main thread's
context.

If you turned off VCR in the main thread after it's context was
duplicated to an additional thread, the additional thread wouldn't see
the new turned off status (since it was only on the linked context), so
it'd fail thinking VCR was still turned on but no cassette was inserted
when a request was made.

Now we check both the current thread's context, and the linked (main)
thread's context to see if VCR is turned off, to ensure an app server
thread can see when VCR is turned off on the main thread.

This was most noticeable in Capybara tests with a JS driver where the
app server thread was started once upon the first spec needing it, thus
duplicating the main thread's context then, but if later specs turned
off VCR, the app server thread never had it's context updated to reflect
that, nor did it check the main thread's context to see if VCR was
turned off there.