Add record_on_error configuration option
When you are TDD'ing an API integration and an error occurs it will
often be caused by an bug in the implementation. In that case no
interactions should be recorded, to avoid having to manually throw away
the cassette every test run until all bugs are solved.

When setting `record_on_error` to false, no interactions will be written
to disk when an error is raised inside the block that uses the cassette.