Support Cucumber-Ruby v4 and later
In version 4 of Cucumber-Ruby the execution of test cases is based on
Pickles. Therefore less data from the feature files are available to
the before/after/etc hooks. Also the Gherkin grammar has changed so
that (all) Scenarios can have Examples - that is "Scenario Outline" is
not just synonym for "Scenario".

Therefore, in case of the "use scenario name" option, use the feature
file name instead of feature name, and use the row of the example to
create uniqueness for Scenarios with Examples.