Improve error message for syntax error in cassette (#909)
When deserializing a cassette and encountering a syntax error, the error
message can be confusing. This is exacerbated in projects using Rails'
backtrace silencer, as the error has no apparent source outside the test
itself.

One common scenario resulting in syntax errors is the use of ERB in the
cassette, but omitting the `:erb` option to `use_cassette`.

For example, a cassette containing may contain a JSON snippet for an API
call where a field is populated using

    "field": <%= "some computed value".to_json %>

If given `erb: true`, this will work fine, but if this is forgotten, a
syntax error will be encountered.

This change appends to the error's message, indicating that the error
was encountered reading a VCR cassette, and suggesting the use of the
`:erb` option.