Make it possible to turn VCR on and off in a new thread
35c5186 fixed so that VCR's `:turned_off` status for the main thread can be read
in a new thread, even if the status changes after the new thread was created.

However, the solution also introduced the problem that the `:turned_off` status
isn't used for other threads than the main thread. This effectively makes it
impossible to turn VCR on and off in a new thread.

This change, instead, creates new contexts with `:turned_off` and
`:ignore_cassettes` set to `nil` by default. When the values are read, they are
only used if they explicitly have been set on the thread's context. Otherwise
the main thread's context is used.